refactor(#2158): split cli_show_security.go by show-security subject
cli_show_security.go had grown to 2018 LOC, just over the ~2000-LOC
modularity threshold (docs/engineering-style.md). Per the converged
#2158 plan (docs/research/2158-file-split/plan.md §5.7) this is a pure
same-package code-motion split: group the `show security` subjects into
sibling files in the same `package cli`.

New sibling files (all package cli):
  - cli_show_security_zones.go    (182): showZonesDisplay
  - cli_show_security_screen.go   (398): showScreen* + screen SYN-cookie rows
  - cli_show_security_objects.go  (291): address-book, applications,
                                         dynamic-address, ALG
  - cli_show_security_ipsec.go    (272): IPsec / IPsec-statistics / IKE
  - cli_show_security_log.go      (232): security log + security alarms
  - cli_show_security_filters.go  (351): firewall filters +
                                         filterTermExpansionCount

cli_show_security.go drops to 342 LOC, keeping the core policies subject
(showPoliciesHitCount, showPoliciesDetail, showMatchPolicies). The
showMatchPolicies doc comment moved with its function; the stray
resolveAddressDetail doc comment stayed in place (its function lives in
cli_show_security_dispatch.go and is still referenced by showPoliciesDetail).

Pure code-motion: functions moved verbatim, no receiver changes, no
package-level var/init (none exist in this file, so no init-order risk).
The only non-move lines are the per-file import blocks and package
declarations. No logic change, no exported-API change.

Validation:
  - go build ./... clean; go test ./pkg/cli/... passes (test set
    byte-identical to master: 606 tests)
  - go doc -all ./pkg/cli/ byte-identical before/after
  - gofmt -l clean on all seven files
  - the pre-existing cli.go:429 "unreachable code" vet note is on master
    and untouched by this change
